// typechain/SvgViewsFacet.d.ts
interface SvgViewsFacetInterface extends ethers.utils.Interface {
  functions: {
    "getAavegotchiSideSvgs(uint256)": FunctionFragment;
    "getItemSvgs(uint256)": FunctionFragment;
    "getItemsSvgs(uint256[])": FunctionFragment;
    "previewSideAavegotchi(uint256,address,int16[6],uint16[16])": FunctionFragment;
    "setSideViewDimensions(tuple[])": FunctionFragment;
    "setSideViewExceptions(tuple[])": FunctionFragment;
  };

  encodeFunctionData(
    functionFragment: "getAavegotchiSideSvgs",
    values: [BigNumberish]
  ): string;
  encodeFunctionData(
    functionFragment: "getItemSvgs",
    values: [BigNumberish]
  ): string;
  encodeFunctionData(
    functionFragment: "getItemsSvgs",
    values: [BigNumberish[]]
  ): string;
  encodeFunctionData(
    functionFragment: "previewSideAavegotchi",
    values: [
      BigNumberish,
      string,
      [
        BigNumberish,
        BigNumberish,
        BigNumberish,
        BigNumberish,
        BigNumberish,
        BigNumberish
      ],
      [
        BigNumberish,
        BigNumberish,
        BigNumberish,
        BigNumberish,
        BigNumberish,
        BigNumberish,
        BigNumberish,
        BigNumberish,
        BigNumberish,
        BigNumberish,
        BigNumberish,
        BigNumberish,
        BigNumberish,
        BigNumberish,
        BigNumberish,
        BigNumberish
      ]
    ]
  ): string;
  encodeFunctionData(
    functionFragment: "setSideViewDimensions",
    values: [
      {
        itemId: BigNumberish;
        side: string;
        dimensions: {
          x: BigNumberish;
          y: BigNumberish;
          width: BigNumberish;
          height: BigNumberish;
        };
      }[]
    ]
  ): string;
  encodeFunctionData(
    functionFragment: "setSideViewExceptions",
    values: [
      {
        itemId: BigNumberish;
        slotPosition: BigNumberish;
        exceptionBool: boolean;
      }[]
    ]
  ): string;

  decodeFunctionResult(
    functionFragment: "getAavegotchiSideSvgs",
    data: BytesLike
  ): Result;
  decodeFunctionResult(
    functionFragment: "getItemSvgs",
    data: BytesLike
  ): Result;
  decodeFunctionResult(
    functionFragment: "getItemsSvgs",
    data: BytesLike
  ): Result;
  decodeFunctionResult(
    functionFragment: "previewSideAavegotchi",
    data: BytesLike
  ): Result;
  decodeFunctionResult(
    functionFragment: "setSideViewDimensions",
    data: BytesLike
  ): Result;
  decodeFunctionResult(
    functionFragment: "setSideViewExceptions",
    data: BytesLike
  ): Result;

  events: {};
}